requirements.txt
Django==5.0.1
Pillow==10.2.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
stripe==7.14.0 